fix: rename support-modules to helpers
