+++ auth/src/helpers/helpers.check-auth-user-index.ts
