+++ auth/src/helpers/helpers.check-auth-user.ts
