+++ auth/src/login/index.ts
@@ -1,6 +1,6 @@
 import { User }           from '../auth.interfaces'
-import { handleError }    from '../support-modules'
-import { checkAuthUser }  from '../support-modules'
+import { handleError }    from '../helpers'
+import { checkAuthUser }  from '../helpers'
 import { compareStrings } from './login.compare-strings'
 
 export const login = (username: string, password: string, usersList: User[]): User[] => {
