+++ auth/src/logout/index.ts
@@ -1,7 +1,7 @@
 import { User }               from '../auth.interfaces'
-import { handleError }        from '../support-modules'
-import { checkAuthUser }      from '../support-modules'
-import { checkAuthUserIndex } from '../support-modules'
+import { handleError }        from '../helpers'
+import { checkAuthUser }      from '../helpers'
+import { checkAuthUserIndex } from '../helpers'
 
 export const logout = (usersList: User[]): User[] => {
   try {
