+++ auth/src/register/index.ts
@@ -1,6 +1,6 @@
 import { User }          from '../auth.interfaces'
-import { handleError }   from '../support-modules'
-import { checkAuthUser } from '../support-modules'
+import { handleError }   from '../helpers'
+import { checkAuthUser } from '../helpers'
 import { hashString }    from './register.hash-string'
 
 export const register = (username: string, password: string, usersList: User[]): User[] => {
