--- auth/src/support-modules/index.ts
@@ -1,3 +1,0 @@
-export * from './support-modules.handle-error'
-export * from './support-modules.check-auth-user'
-export * from './support-modules.check-auth-user-index'
